requirements.txt
Adafruit-Blinka==8.47.0
Adafruit-BME280==1.0.1
adafruit-circuitpython-bme280==2.6.25
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-requests==4.1.6
adafruit-circuitpython-typing==1.11.0
Adafruit-PlatformDetect==3.74.0
Adafruit-PureIO==1.1.11
Adafruit_GPIO==1.0.3
binho-host-adapter==0.1.6
blinker==1.8.2
board==1.0
click==8.1.7
colorzero==2.0
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
fonttools==4.54.1
gpiozero==2.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
lgpio==0.2.2.0
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pyftdi==0.55.4
pyparsing==3.1.4
pyserial==3.5
PySide6==6.7.3
PySide6_Addons==6.7.3
PySide6_Essentials==6.7.3
python-dateutil==2.9.0.post0
pytz==2024.2
pyusb==1.2.1
rpi-lgpio==0.6
scipy==1.14.1
setuptools==75.1.0
shiboken6==6.7.3
six==1.16.0
spidev==3.6
sysv_ipc==1.1.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.4
